requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
buildozer==1.5.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
distlib==0.3.9
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.44.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
ptyprocess==0.7.0
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sh==2.2.2
six==1.17.0
smmap==5.0.2
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
watchdog==6.0.0
